requirements.txt
pyusb==1.2.1
pyside6==6.7.0
